[kernel][arm][mmu] skip break-before-make during early kernel boot
This allows the kernel to split its own physmap and its own kernel
mappings' large pages while the system is still running on the bootup
core.

Break-before-make causes some issues during early kernel boot,
specifically when the kernel unmaps sections of the physmap and
re-protects sections of the kernel image itself. In these cases based on
how the kernel address space was set up initially it may need to split
pages. In various scenarios this could result in either temporarily
unmapping the kernel itself while its running or unmapping a section of
the physmap that the mmu code needs to access the page tables.

For now, continuing to allow break-before-make gets the kernel over this
hump, though it's technically pushing the architecture a bit, since not
BBM when splitting a large page *may* cause a TLB conflict abort,
depending on the core implementation. There's no reason to believe any
cores that exist exhibit this behavior, but it may be a thing in the
future.

Change-Id: I58c6643d4c02f2751fc0cced1d03bac847e70da6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1039220
Commit-Queue: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Reviewed-by: Will Drewry <drewry@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>