[shac] Use shared shac checks in fuchsia.git.
This modifies the starlark.star file to use the buildifier format
check from the shared repository. There should be no change in
behavior.

This also registers the buildifier_lint check from the shared repo
configured as a warning. There are 32 files with linting issues
across the tree that I will address before enabling this
as blocking. (or if that is undesirable, I will try again to land a shac.star file in the //build/bazel_sdk directory)

This also also makes it possible to run the clang-format check from
the shared repository, though it results in an error (b/b/338231347)

Bug: b/338231347
Change-Id: I3e92aead2ba036b7e6fbe2158d82fb4a40566357
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1040164
Commit-Queue: Jerry Belton <jcecil@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>