[sysmem2] fuchsia-framebuffer to sysmem2 done
This finishes the move of fuchsia-framebuffer to sysmem2.

By renaming sysmem2.rs to sysmem.rs (instead of deleting sysmem.rs),
the line-by-line git info for sysmem.rs is preserved, and we don't
have to have the "2" in client code now that they're all moved to
sysmem2.

BUG: b/306258175
Change-Id: I436b3f182500e89e7b58692a38c30e9f1249bdf0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1032396
Reviewed-by: Sergey Ulanov <sergeyu@google.com>
Commit-Queue: Dustin Green <dustingreen@google.com>
Reviewed-by: Emircan Uysaler <emircan@google.com>
Reviewed-by: Will Drewry <drewry@google.com>