[kernel][vm] Remove extraneous print
This print generates unnecessary spurious output, including when
running the kernel unittests. The correct operation when this scenario
occurs, a pinned page in a discardable VMO, is checked programatically
in the unittests and does not require this print to validate.

Change-Id: Ic458e2c44885c0480416150a26ad597c66ce7bea
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1041874
Fuchsia-Auto-Submit: Adrian Danis <adanis@google.com>
Reviewed-by: David Worsham <dworsham@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>