[netstack3] Apply selected multicast interface on connect()
Previously selected multicast interface in SendOptions was applied only
for sendto(). It should be applied when connecting the socket in order
to select the correct route and bind the socket to the correct local
IP. This CL removes the option from SendOptions. Instead it's applied
for datagram sockets from DatagramSocketOptions.
Tests for this functionality are added in follow-up CLs.

Bug: 331651296
Change-Id: I20f2e162dd0f574e639ff7d5e440abc76a99323e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1041946
Commit-Queue: Sergey Ulanov <sergeyu@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>