[fxtest] Build updates for e2e tests
Running e2e tests under fx test often underbuilds. This is because e2e
tests depend on components running globally on the system, but updates
to these components are not reflected in the build of just the host-side
script for the test. This CL ensures that if an e2e test is selected,
the "updates" package is rebuilt so that a subsequent publish will
result in dependencies being updated.

Note that this does not restart any components that may be running.

Fixed: 305282664
Change-Id: I19c03c0dac463f7d52f30505fffe37d4220ff11f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1043377
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Fuchsia-Auto-Submit: Christopher Johnson <crjohns@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>