[ffx] Add more logging to daemon start-up
There is sometimes a hang in the daemon start-up. The last message
produced in ffx.daemon.log is "ffx_daemon_start: creating daemon", and
we're not seeing the next expected message: "starting daemon".  This
adds a couple more log messages to we can start bisecting where the
daemon is hanging.

Bug: b/316626057
Bug: b/333744245
Change-Id: Ice3345195f8534c0363a6b1400e054560b3a9bf2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1043243
Reviewed-by: Andrew Davies <awdavies@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>