[fxfs] Add ToWeakNode
It is possible for there to be one other existing strong reference which
is dropped after the `upgrade()` call here but before the local strong
reference is dropped for not being an FxFile. If that happens we will be
dropping the last strong reference inside the lock which will trigger
the `Drop()` on the Node which will try to remove it from the cache
resulting in a deadlock.

Bug: b/338634316
Change-Id: I28b02694547a460564cad26aa43985c1776bdc15
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1043612
Commit-Queue: Chris Suter <csuter@google.com>
Reviewed-by: Martin Lindsay <mlindsay@google.com>