[component_manager] actions don't need to ensure Discover runs first
When `ResolvedInstanceState::new` populates the `children` hashmap with
newly created components, the function discovers each child before
returning. This means that all children are discovered before the
instance state mutex of the parent component is released, ensuring that
all static children are discovered before it's possible to get a
reference to them. Unfortunately this wasn't true for dynamic children
or the root component.

This commit rearranges the codepath for adding dynamic children such
that dynamic children are discovered before references to them become
accessible, along with the codepath by which the component model is
started to ensure that the root component is discovered before
capabilities accessible through component manager's out directory start
running.

With these changes action implementations do not need to be careful
about ensuring that they wait for a discover action to run, as it's safe
to assume that if a reference can be obtained to a component then the
component has been discovered.

Multiply: namespace-capabilities-integration-test
Multiply: controller_integration_test
Multiply: stop_with_escrow_integration_test
Change-Id: Id89b0fa8517ce312b3574cfcc20d3df30082d178
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1042093
Commit-Queue: Claire Gonyeo <cgonyeo@google.com>
Reviewed-by: Gary Bressler <geb@google.com>