[component_manager] removed Actions::finish
finish should not be part of the public API of the Actions struct.

Prior to this change ActionSet relied on being able to reach itself and
call finish through `ComponentInstance::lock_actions`, which is no
longer possible. To work around this, ActionSet is now held in an
Arc<Mutex<_>> so that it can hold a reference to itself and lock itself
once an action finishes running.

Change-Id: Iff68786aaae0982ad661539a2c0094f43c489e7d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1041374
Reviewed-by: Gary Bressler <geb@google.com>
Fuchsia-Auto-Submit: Claire Gonyeo <cgonyeo@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>