[vim3][display] Add a display detection driver
Prior to this CL, the bootloader passed in driver.vim3.has_lcd variable
which was read to figure out if HDMI or MIPI_DSI display was connected.
Ideally we wanted to determine this state by probing the LCD_RESET gpio
in fuchsia itself. This CL adds a vim3-display-detect driver that does
the same and adds a display child node with the correct output type
property. With the help of bind rules, the correct drivers get started.

Bug: 42076008
Test: Booted core.vim3 with Khadas TS-050 DSI display
Change-Id: I53ff4773b9db3be0df1f1e99919afc10460a6381
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1043544
Reviewed-by: Yilong Li <liyl@google.com>
Fuchsia-Auto-Submit: Victor Costan <costan@google.com>
Reviewed-by: Jocelyn Dang <jocelyndang@google.com>
Reviewed-by: Victor Costan <costan@google.com>
Commit-Queue: Victor Costan <costan@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>