[uart] Fix blocking option not being surfaced into the waiter.
There are two flavors of dputs one that will never block, and one
that may block. I mistakenly mapped may block to the case were it would
not block when debuglog is disabled.

Note that the splitting of the flavors is correct, just the mapping for
the newer drivers was incorrect. The Waiter needs to, first check which
flavor it is, and then check whether its allowed to block.

The fix is just forwarding a template parameter to Waiter from the Write
call.

Fixes: 336859048
Change-Id: Idc80d78041d35eacc79e737cf3d723078a51fe83
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1036197
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>