[snapshot] Enhance e2e test to also validate log files and others
For PETS purposes we can likely reuse this test to ensure logging and
inspect are working by validating the presence of data in a snapshot.

We can likely reuse this test by just enhancing it to not only validate
the inspect.json file, but also the log files as well as other metadata
files present in the zip.

Multiply: snapshot_test_fc
Bug: 338416932
Change-Id: Ife4b43b48c39c57c4a70eb35de88adc7187d63e2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1040853
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Alex Pankhurst <pankhurst@google.com>
Commit-Queue: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Kevin Cho <chok@google.com>