[netstack3] Use raw parsing for transport header
Currently, when dealing with an incoming packet, the filtering crate
fully parses the transport header in order to extract source and
destination ports for inspection. This ends up doing a lot of
unnecessary work, in particular to verify the transport layer checksum,
which can be very expensive when done repeatedly on all packets (as
happens if there are any ingress filtering rules that match on
transport-layer fields).

Instead of doing full parsing, use the "raw parsing" facilities from
packet-formats to only parse the fixed portion of the transport header,
which is enough to extract the ports, and is much cheaper.

In the future, we could consider caching the result of this parsing (for
example, by storing it as a field on `RxPacket` and introducing a
wrapper type for `Ipv{4,6}Packet` that similarly stores it) so that raw
parsing is not repeated for every call to IpPacket::transport_packet.

Test: netstack3-filter-test
      netstack-filter-fidl-tests
Change-Id: Ie7449089abf2d88710419910b8d289864c690fe8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1044977
Commit-Queue: Peter Johnston <peterjohnston@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Peter DeLong <pdelong@google.com>