[netstack3] Rename FakeEventDispatcherBuilder
`EventDispatcher` is a very old name for the bindings contexts that is
no longer used, this is an artifact from a distant past.

The builder is still used to create `FakeCtx`s for tests, so renamed it
to `FakeCtxBuilder` instead.

Bug: 338448790
Change-Id: Ide9419784c9acf2af2a412dff4f187c4b60c1fc4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1045596
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Evan Wildenhain <wildenhain@google.com>