[power][sag] Create FakeSAG AIB
This change splits out the system-activity-governor out of the power
framework shard, so that we can have two variants of it based on which
AIB gets picked from the configuration. The testing variant is the new
fake SAG, and we also allow it to be accessed from the console launcher
so that bootfs tests can access its protocols. This change also allows
power framework to be included in the bootstrap feature set. Finally it
allows the fake-suspend driver to bind to the same node the real suspend
driver can bind to.

Bug: 335526421
Change-Id: I663fc1ab2b8bd316dd3015b68549ed7a50aba796
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1042045
Reviewed-by: Aidan Wolter <awolter@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Fuchsia-Auto-Submit: Novin Changizi <novinc@google.com>
Reviewed-by: Didi She <didis@google.com>
Commit-Queue: Novin Changizi <novinc@google.com>