[build][assembly] Refactor of AIBs to remove 'main' compiled component definition
This finally removes the CompiledPackageMainDefinition and 'main' vs
additional-shards distinction.  It also moves tracking of the input
cml files to the GN template for AIBs, so that cml changes are known
to trigger a re-building of the AIBs (although they should have been
tracked via depfiles before this change).

This is in preparation for using the same patterns to add compiled
package/component support to developer overrides.

Run-All-Tests: true

Change-Id: I71fa4f7ac3c0fa04c7acbc0099ba59f8864bf683
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1044452
Reviewed-by: Aidan Wolter <awolter@google.com>
Commit-Queue: Aaron Wood <aaronwood@google.com>