[testing] Ensure gtest runner reports open failures
When opening a file without a describe event, failures will not be
noticed so the channel may be closed by the time we try to read from
the file. In cases where the file doesn't exist, for example, now we
will report the correct failure `NOT_FOUND` instead of a generic
`PEER_CLOSED` error lacking context about the failure.

Test: No functional change.

Change-Id: I1bd11f3aa24e158008d9caaea3eea8d3a4f209af
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1045273
Reviewed-by: Ankur Mittal <anmittal@google.com>
Commit-Queue: Brandon Castellano <bcastell@google.com>