[component_manager] Fix shutdown panic bug
It's possible to cause shutdown to panic by constructing a cycle with
dynamic offers. Specifically, this is possible if there is a service
offer from a collection, followed by a chain of dynamic offers that
routes back to this collection.

To make this impossible, detect these cycles at component creation time.

Fixed: 297403341
Change-Id: I8ffa6636083b7364ea4d55201a9e7fa84eb92aa1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1030252
Commit-Queue: Gary Bressler <geb@google.com>
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>