[ffx] Support ffx --log-output <file/stdio> ...
Add an option to direct the output to a particular destination.
If the destination file is "-" or "stdout", send the output to stdout.
If it is "stderr", send to stderr.

This reinstates the functionality (lost in fxr/1042050) that daemon
output should go into ffx.daemon.log instead. (It also makes it so
that `ffx daemon start` produces output on stdout, retaining
compatibility, with the previous behavior, since this is necessary
for integrators who manage the daemon lifecycles themselves.)

Bug: b/338451284
Change-Id: I3d02b3a63423e8ff0cc7bbf137c3f69373de9311
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1045226
Commit-Queue: Steven Grady <slgrady@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>