[starnix] Add ffx command to suspend container
This command currently suspends all running Starnix containers by
suspending each process in the container.

It does not suspend processes that may be logically associated with a
container if they are not in the same job.

In order to make it simple to suspend a single container, we could add
suspend functionality to fuchsia.starnix.container.Controller. This
would allow clients to connect to the container they wanted to suspend
directly. However, the suspend operation would still need to be
performed by a separate component, outside of the container's job. This
means that we would need to rework the division of labor between the
starnix_kernel and starnix_runner (mainly which component serves protocols
on behalf of the container).

Change-Id: I4a3c3b2e6cfe3df575466cd55434d8e2e74cb835
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1043153
Reviewed-by: Adam Barth <abarth@google.com>
API-Review: Adam Barth <abarth@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>
Commit-Queue: Kevin Lindkvist <lindkvist@google.com>