[superproject] Roll third_party/mesa [lavapipe/fuchsia] Generated file for buffer collection ext.
The buffer collection extension is contained in a separate source
set which didn't explicitly depend on the target which generates
`lvp_endpoints.h`.  This results in a race condition in the build,
where sometimes the file isn't available when it is needed.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/third_party/mesa/+/1047313
Original-Revision: 47bc656ecd6cc432ec5ba5d465ab3a8f2b8ce649
GitOrigin-RevId: 2a7e6bb433afd28aee3952a42e2204aad3ca2f4b
Roller-URL: https://ci.chromium.org/b/8748002688886885361
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: Ibf3af92cb17de9852043441806c918697dc8c0ff
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1047264