[netstack/conntrack] Only collect connections once
By only checking if a connection is expired when we're looking at it
through the original tuple, we cut the size of to_remove in half. It
also means we can add in asserts that we actually did remove the
connections from the table.

Change-Id: Icc72186afeae28df8ee8f7d19922cea815c2e731
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1046619
Commit-Queue: Peter DeLong <pdelong@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>