[packet-formats] No pseudo header in ICMPv4 checksum
ICMPv4 packets produced by packet-formats do not include the IP pseudo
header in their transport-layer checksum calculation, because the IPv4
header always includes a checksum. Fix the pseudo header checksum update
logic to correctly do nothing for ICMPv4.

The existing test did not catch this because it was swapping the source
and destination IP addresses, resulting in a checksum that was the same
as if those addresses had never been included. Fix the test to use
different IP addresses for both source and destination such that it
catches this bug.

Bug: 321013529
Test: netstack3-filter-test
Change-Id: I845e0af2bbe73d1542bf4e2c0f5bc4ab7eb38aeb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1048072
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Peter DeLong <pdelong@google.com>