[bazel][bazel_sdk] Work around for Bazel analysis bug.
Bazel has a subtle bug when aspects are mixed with dependency
paths between testonly and non-testonly targets. More precisely,
if a testonly target depends on a non-testonly one, Bazel injects
some undocumented TestTrimmingTransition transition calls between
them at analysis time.

This is supposed to remove the test-specific fragment from the
build configuration when reaching the non-testonly dependency.

*However*, this magic does not happen when aspects are used to
walk the same dependency chain.

This results in two instances of the same cc_library() being
created in the context of two different build configurations,
which only differ in values that do not affect the generated
actions _nor_ the location of their output artifacts.

Bazel then complains that two different actions are trying
to write to the same output file (e.g. `..../lib.pic.o`).

The only way to disable that is to use a specific command-line
option, which can be either [1] or [2], where the former is
considered less risky.

This CL ensures the flag is used in the .bazelrc files of
the workspaces for the Fuchsia platform build, and of the
Bazel SDK test suite.

OOT projects will have to add this option to their own
.bazelrc file, unfortunately :-/

+ Add a new Bazel SDK test that builds a Fuchsia test
  package that includes a test component using a
  fuchsia_cc_test() binary that depends on a non-testonly
  cc_library().

  For now, this test will pass even if the flag is not added
  to the .bazelrc, but this will no longer true in a future
  CL that changes the Bazel SDK rules logic to use aspects
  to collect unstripped binaries.

+ Add proper @com_google_googletest override when invoking
  Bazel SDK test suite with bazel_test.py script from the
  platform build.

[1] https://bazel.build/reference/command-line-reference#flag--experimental_retain_test_configuration_across_testonly
[2] https://bazel.build/reference/command-line-reference#flag--trim_test_configuration

Change-Id: Ic9d27734937248c17afff6ba9fabceace956d6a8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1047976
Reviewed-by: Chase Latta <chaselatta@google.com>
Commit-Queue: David Turner <digit@google.com>