[superproject] Roll llvm-project/libc [libc] Fix GPU handling for unsupported backends (#92271)
Summary:
If the user does not have the selected backend enabled, we should still
be able to build the LLVM-IR an ddistribute it. This patch makes logic
to suppress tests if the backend can't build it, as well as removing a
flag for the building that's only present int he NVPTX backend.
GitOrigin-RevId: e5a190cf1567a4a62e0bcc6d79dcdaa5a9866308
Original-Revision: 64cf380a67fa8e8c844d52e4d1f22b9db5fe9832
Roller-URL: https://ci.chromium.org/b/8747832895693508705
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: Ia4a3450fbfb31329f5c8920ac592f6bf0eae4ee2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1048534