[block][sdhci] Refactor PendingRequest
Move PendingRequest into an optional. This will be helpful when buffered
IO is implemented, as it enables the lifetime of an fzl::VmoMapper to be
tied to the PendingRequest object.

Also simplify the PendingRequest status fields somewhat, and use them to
enforce the expected interrupt order (i.e. it is acceptable for the
command and transfer complete interrupts to be received at the same
time, but not for the latter to be received before the former).

Test: sdhci-test
Change-Id: I2d77f2037bc66011095feee8d8b0162d17646ea8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1048376
Commit-Queue: Braden Kell <bradenkell@google.com>
Reviewed-by: HanBin Yoon <hanbinyoon@google.com>