[kernel][vm] Remove page request batching
The batching system, both internal and external, are vestigial and from
a time when every page offset was looked up individually for a request.
Current lookup cursors and methods now find and request their entire
range at once, and do not need to append to an existing range. The
current batching logic is, therefore, completely unused and the code
paths are not triggered.

Although batching itself is unused, a page request having been created
with a different batch state changed how failed ranges worked. This
behavior does not make sense, and the comment justifying it is wrong,
since the external callers are in fact explicitly requesting the range
that is being failed, and so will not be surprised. Changing this
behavior does impact some tests that were explicitly testing for it.

Change-Id: I5be09b9baf73080bf848a5ea7b2f7374c216de68
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1045573
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: David Worsham <dworsham@google.com>