[bazel_sdk] Use aspect to collect ELF binaries when building Fuchsia packages.
Use the find_and_process_unstripped_binaries() rule to collect
all FuchsiaUnstrippedBinaryInfo providers from a DAG of dependencies,
and use the result to package the corresponding stripped binaries
into a Fuchsia package, and expose the unstripped symbol libraries
to the debugger.

To do this:

- _fuchsia_cc() now exposes the generated (unstripped) binary with
  a FuchsiaUnstrippedBinaryInfo provider (and does not add a
  corresponding FuchsiaPackageResourcesInfo entry).

- The fuchsia_package() and fuchsia_test_package() macros now create
  a new target using find_and_process_unstripped_binaries(), and
  pass it as a dependency to the underlying _build_fuchsia_package()
  or _build_fuchsia_test_package() rule.

  Said rule will extract the FuchsiaPackageResourcesInfo from the
  target (which lists stripped binaries and their install location)
  and also its FuchsiaDebugSymbolInfo value (which points to the
  corresponding .build-id/ directory).

This should not change the result, only the way dependencies
are collected.

Note that this doesn't remove the logic inside
_build_fuchsia_package_impl() that tries to strip _all_ resources
listed for the package. This code may be removed later when the
SDK atoms for prebuilt libraries also use a
fuchsia_unstripped_binary() rule to expose their content.

This also simplifies the fuchsia_driver_tool() rule which
can now directly get the path of the tool binary without
scanning resource paths or requiring a `tool_entry_point`
attribute.

Bug: 339038603
Change-Id: I842efd3cd3299f93dbd6e7726817a7a314ade489
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1047133
Reviewed-by: Darren Chan <chandarren@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>
Commit-Queue: David Turner <digit@google.com>