[starnix/usercopy] Introduce ASM `hermetic_zero`
Currently, we build `hermetic_zero` from C++ source, but the hermetic
zero error path assumes certain behaviour of the `hermetic_zero`
implementation (see `hermetic_copy_error`. Compiling from C++ source
is brittle because the generated assembly may change across toolchain
updates. In fact, we have observed breakages several times.

Now that the `hermetic_zero` is stable and we are not iterating on it
anymore, provide an assembly variant of it that we can build from so
that future toolchain updates do not break the build and, most
importantly, we can explicitly provide the hermetic zero error path
the guarantees/behaviours it expects.

Note that we keep the `hermetic_zero` C++ source in tree as a reference.
A later change will remove its `hermetic_code_blob` target from the
regular build graph and likely move it to a test-only build target.

Bug: b/338636637
Change-Id: I740c89e584f703113c7ed88b86304f91abace86d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1045755
Reviewed-by: James Robinson <jamesr@google.com>
Commit-Queue: Ghanan Gowripalan <ghanan@google.com>