[cmc] Clear source_dictionary when source is voided
Fix a bug where, if `source_availability: unknown` is used and the
source is a dictionary exposed by a child, the source is set to Void but
the source_dictionary is left set.

Not clearing source_dictionary may break assumptions about state (in the
validator for example).

Fixed: 340624184

Change-Id: I7a82793dc4790d7f53a0ce0f9c29333c8de58898
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1048374
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Gary Bressler <geb@google.com>