[bt][a2dp] Fix compatible endpoints not matching
Update the logic for matching compatible codecs (during start) to
include all endpoints which we can negotiate a compatible set of codecs
for.  We continue to be stringent when setting our configuration.

Add negotiation to the start process to use the highest quality codec
parameters that are compatible.

Fixed: 339159944
Test: added unit tests
Test: manual test with previously failing headset
Change-Id: I22356168ab891cb4260d077561bbf1e86ff82e82
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1048244
Commit-Queue: Marie Janssen <jamuraa@google.com>
Reviewed-by: Ani Ramakrishnan <aniramakri@google.com>