[vim3-devicetree] Fix network connectivity issue
There were two issues which was leading for the vim3-devicetree based
devices to be unreachable. One was the netsvc_interface path, which
differs between vim3 and vim3-devicetree (on vim3-devicetree, the path
name is derived from the node name and changing that to match that of
vim3 is a major change). The other issue was passing of the serial
number metadata was missing for usb peripheral on vim3 devicetree.

Fixed: 340578181
Test: Checked that the device was accessible through fx shell.
Change-Id: Ie3ed6bc63fce2f4967928027924aee29c4d7a532
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1048094
Reviewed-by: Braden Kell <bradenkell@google.com>
Reviewed-by: James Robinson <jamesr@google.com>
Fuchsia-Auto-Submit: Puneetha Ramachandra <puneetha@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>