[ffx] Move HostPipeErr to SshError
While refactoring to improve determining when SSH error are fatal or
should have a retry, it became clear that using HostPipeErr seemed
strange. Its location doesn't make sense, and it creates circular
dependencies when used in the ffx `target` library.

This refactors the error into a more reasonable location, and also
renames it to make it clearer that this is making an assumption about
underlying implementation rather than an abstraction.

Change-Id: Idf9c1d4795d760e62aeaa064db4d63a03d21ebfb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1050052
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Andrew Davies <awdavies@google.com>