[kernel_zbi_test] Propagate device_type to cuckoo test
The test should only be executed in the device types set in the
zbi_test. This were not forwarded since it expected environment,
but boot_test expects device_types. So we translate device types to
environments.

Reenable cuckoo test, but given it will not be executed outside of nuc7,
and coverage is only QEMU bots for the time being, then the overall
effect is the same.

A separate issue is HPET reliability in QEMU.

Change-Id: I3329c0e69cb141e180ccf095ad533a363926307e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1049718
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>