[fsverity] Set the digest_algorithm/size output fields in measure_verity
In its old implementation, fsverity treats the digest_algorithm field of
the fsverity_digest struct as an input field but it is actually only an
output field. This change sets both the digest_algorithm and digest_size
of the fsverity_digest struct that is returned to the user. This should
also fix the begonia boot regression bug attached to this CL, which was
caused by the EINVAL error returned from a check that treated the
digest_algorithm as an input field.

Bug: b/340066061
Change-Id: I5f035b739bf07181f08ed481e8b816d9b0a5ab30
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1049193
Reviewed-by: James Robinson <jamesr@google.com>
Commit-Queue: Nikita Jindal <nikitajindal@google.com>