[netstack3] Introduce SocketZonedAddrExt
Instead of a freestanding function in the transport module, move
resolve_addr_with_device to an extension trait on ZonedAddr so we don't
need to import freestanding functions, and define it in the socket
module which will be extracted to the base crate.

This allows TCP and UDP to use the extension directly from the base
crate without the need for a base transport module.

In preparation for moving socket, move EitherDeviceId to the base crate
so the new extension trait can reference it.

Bug: 338448926
Change-Id: Ib94aa4163d62062677e7c8b836d4bc39a1a8d296
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1050599
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>