[netstack3] Add TransportPacketMut::set_{src,dst}_port
Add methods that allow changing the source and destination port in the
transport header of an IP packet. For incoming packets this requires
updating the checksum in place as well.

Note that for ICMP packets, it continues to be the case that only echo
requests and replies support transport-layer inspection and
modification. Just as an ICMP echo request encodes its ID as the source
port and a reply encodes its ID as the destination port, only those same
ports can be rewritten.

Fixed: 321013529
Test: netstack3-filter-test
      netstack3-core-test
Change-Id: I766d76a930d86b36ee0e89d2efcaec0e82747ccb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1046012
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Peter DeLong <pdelong@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>